refactor: move AnimationPlayback module from stores to models directory
